fasthtml-example/requirements.txt
python-fasthtml
plotly
fh-plotly